% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_icd_labels}
\alias{all_icd_labels}
\title{ALL ICD9 & ICD10 Labels}
\format{
A data frame with 113,272 rows and 3 variables:
\describe{
  \item{dx}{ICD-9-CM or ICD-10-CM code}
  \item{dx_ver}{Indicator of ICD-9 or ICD-10}
  \item{desc}{Code description}
}
}
\source{
\url{https://www.cms.gov/medicare/icd-10/2022-icd-10-cm}
}
\usage{
all_icd_labels
}
\description{
A dataset containing labels for all ICD9 & ICD10 codes
}
\keyword{datasets}
