% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{avail_disease_codes}
\alias{avail_disease_codes}
\title{Return available disease code sets}
\usage{
avail_disease_codes(description = TRUE)
}
\arguments{
\item{description}{a logical indicator for whether to return descriptions}
}
\description{
Return the names and descriptions of the available disease code sets
}
