% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{avail_misc_codes}
\alias{avail_misc_codes}
\title{Return available miscellaneous code sets}
\usage{
avail_misc_codes(description = TRUE)
}
\arguments{
\item{description}{a logical indicator for whether to return descriptions}
}
\description{
Return the names and descriptions of the available miscellaneous code sets
}
