% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{avail_procedure_codes}
\alias{avail_procedure_codes}
\title{Return available procedure code sets}
\usage{
avail_procedure_codes(description = TRUE)
}
\arguments{
\item{description}{a logical indicator for whether to return descriptions}
}
\description{
Return the names and descriptions of the available procedure code sets
}
