% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{avail_rx_codes}
\alias{avail_rx_codes}
\title{Return available rx code sets}
\usage{
avail_rx_codes(description = TRUE)
}
\arguments{
\item{description}{a logical indicator for whether to return descriptions}
}
\description{
Return the names and descriptions of the available disease code sets
}
