% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{avail_symptom_codes}
\alias{avail_symptom_codes}
\title{Return available symptom code sets}
\usage{
avail_symptom_codes(description = TRUE)
}
\arguments{
\item{description}{a logical indicator for whether to return descriptions}
}
\description{
Return the names and descriptions of the available symptom code sets
}
