% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccs10_mappings}
\alias{ccs10_mappings}
\title{Clinical Classification Software: ICD-10-CM Diagnosis CCS codes}
\format{
A data frame with 72,446 rows and 8 variables:
\describe{
  \item{icd10cm}{ICD-10-CM diagnosis code}
  \item{ccs_code}{CCS code number}
  \item{icd_description}{ICD-10-CM Description}
  \item{ccs_description}{CCS Code Description}
  \item{ccs_multi_lvl_1_code}{CCS Multi-level Level 1 Code}
  \item{ccs_multi_lvl_1_description}{CCS Multi-level Level 1 Description}
  \item{ccs_multi_lvl_2_code}{CCS Multi-level Level 2 Code}
  \item{ccs_multi_lvl_2_description}{CCS Multi-level Level 2 Description}
}
}
\source{
Version: CCS v2019.1 \url{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/ccsr_archive.jsp#ccsr}
}
\usage{
ccs10_mappings
}
\description{
A dataset containing CCS codes for ICD-10 diagnoses
}
\keyword{datasets}
