% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccs9_mappings}
\alias{ccs9_mappings}
\title{Clinical Classification Software: ICD-9-CM Diagnosis CCS codes}
\format{
A data frame with 15,073 rows and 4 variables:
\describe{
  \item{icd9cm}{ICD-9-CM diagnosis code}
  \item{ccs_code}{CCS code number}
  \item{ccs_description}{CCS Code Description}
  \item{icd_description}{ICD-10-CM Description}
}
}
\source{
Version: 2015 \url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/ccs.jsp}
}
\usage{
ccs9_mappings
}
\description{
A dataset containing CCS codes for ICD-9 diagnoses
}
\keyword{datasets}
