% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccs_labels}
\alias{ccs_labels}
\title{Clinical Classification Software: CCS Diagnosis Code Labels}
\format{
A data frame with 285 rows and 2 variables:
\describe{
  \item{ccs_code}{CCS code number}
  \item{ccs_category}{CCS Code Group Description}
}
}
\source{
Version: 2015 \url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/ccs.jsp}
}
\usage{
ccs_labels
}
\description{
A dataset containing labels for CCS diagnosis groups
}
\keyword{datasets}
