% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_icd_functions.R
\name{children10}
\alias{children10}
\title{Get 10 Child codes}
\usage{
children10(icd_10_codes, only_code = TRUE)
}
\arguments{
\item{icd_codes}{a vector of icd codes to get children for}
}
\description{
Returns the set of child codes corresponding to a given set of ICD 10 codes. 
This updates the children function from the ICD package to use the most recent
library of ICD codes.
}
