% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_icd_functions.R
\name{children_safe}
\alias{children_safe}
\title{Get ICD-9/10 Child codes (safe version)}
\usage{
children_safe(icd_codes)
}
\arguments{
\item{icd_codes}{a vector of icd codes to get children for}
}
\description{
Returns the set of child codes corresponding to a given set of icd_codes.
This is a safer version than that of the icd package, as it retains a given
code even if no child codes can be found.
}
