% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_icd_functions.R
\name{describe_icd10}
\alias{describe_icd10}
\title{Describe ICD-10 codes}
\usage{
describe_icd10(icd_codes, return_short = FALSE)
}
\arguments{
\item{icd_codes}{a vector of ICD-10 codes to get descriptions.}

\item{return_short}{an indicator of whether to only return the short code description
as oppoed to the long description (default is FALSE)}
}
\description{
Returns the labels for a set of ICD-10 codes.
}
