% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_funcs.R
\name{find_ccs}
\alias{find_ccs}
\title{Find CCS category containing a given term}
\usage{
find_ccs(term)
}
\arguments{
\item{term}{A term (string) to search within CCS categories}
}
\description{
`explain_ccs` returns a tibble of CCS codes and category labels
where the category contains a given term
}
\examples{
find_ccs("flu")
find_ccs("fever")
}
