% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{find_tokens}
\alias{find_tokens}
\title{Find tokens corresponding to diseases and symptoms}
\usage{
find_tokens(term_list)
}
\arguments{
\item{term_list}{a character vector of terms to scan for token matches}
}
\description{
Return the tokens and corresponding disease and symptom sets from a list of
terms. Each term input is matched to all potential tokens
}
