% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook_funcs.R
\name{get_ndc_from_terms}
\alias{get_ndc_from_terms}
\title{Get NDC code from medication name}
\usage{
get_ndc_from_terms(
  terms,
  return_all = FALSE,
  redbook_path = "~/Data/redbook.csv"
)
}
\arguments{
\item{terms}{A vector of medication names}

\item{return_all}{A logical indicator of whether to return all values. Default
value is set to FALSE (if FALSE only ndc codes are returned)}

\item{redbook_path}{The path to the redbook file.}
}
\description{
`get_ndc_from_name` returns the ndc code, or the corresponding redbook
 entries, for medications of a given name. The name provided is matched to
 the variables "genme" and "prodnme" in the redbook file.
}
