% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_map_ahrq_corrected}
\alias{icd10_map_ahrq_corrected}
\title{AHRQ Elixhauser Comorbidity Score Indicators for ICD-10 Codes}
\format{
A list with 30 vectors named for each
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp#overview}
}
\usage{
icd10_map_ahrq_corrected
}
\description{
A dataset containing corrected indicator lists of ICD-10 codes for each comorbidity.
Codes were taken from software lists from 2016-2022. Details of construction
can be found in the comorbidity repository.
}
\keyword{datasets}
