% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_funcs.R
\name{icd10_to_icd9}
\alias{icd10_to_icd9}
\title{Convert ICD-10-CM values to ICD-9-10}
\usage{
icd10_to_icd9(icd_list)
}
\arguments{
\item{icd_list}{A vector of ICD-10 values}
}
\description{
`icd10_to_icd9` converts a vector of ICD-10 values to possible ICD-9 values.
Uses the NBER crosswalk file.
}
\examples{
icd10_to_icd9("A047")
}
