% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_labels}
\alias{icd10cm_labels}
\title{ICD10 Labels}
\format{
A data frame with 95,711 rows and 4 variables:
\describe{
  \item{code}{ICD-10-CM code}
  \item{billable}{Indicator if the code is a billable code}
  \item{short_desc}{Short code description}
  \item{long_desc}{Long code description}
}
}
\source{
\url{https://www.cms.gov/medicare/icd-10/2022-icd-10-cm}
}
\usage{
icd10cm_labels
}
\description{
A dataset containing ICD-10-CM codes along with short and long descriptions. This
dataset was created from the base data inside the icd package from `icd::icd10cm2019`
and appended with CMS order labels from 2017-2022
}
\keyword{datasets}
