% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_to_9}
\alias{icd10cm_to_9}
\title{NBER Diagnosis Crosswalk: ICD-10-CM to ICD-9-CM}
\format{
A data frame with 78,838 rows and 8 variables:
\describe{
  \item{icd10cm}{ICD-10-CM code}
  \item{icd9cm}{ICD-9-CM code}
  \item{flags}{(Describe "flags")}
  \item{approximate}{(Describe "approximate")}
  \item{no_map}{(Describe "no_map")}
  \item{combination}{(Describe "combination")}
  \item{scenario}{(Describe "scenario")}
  \item{choice_list}{(Describe "choice_list")}
}
}
\source{
\url{https://www.nber.org/research/data/icd-9-cm-and-icd-10-cm-and-icd-10-pcs-crosswalk-or-general-equivalence-mappings}
}
\usage{
icd10cm_to_9
}
\description{
A dataset containing crosswalk values from ICD-10 diagnosis codes to ICD-9
diagnosis codes
}
\keyword{datasets}
