% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10pcs_to_9}
\alias{icd10pcs_to_9}
\title{NBER Procedure Crosswalk: ICD-10-PCS to ICD-9-PCS}
\format{
A data frame with 82,571 rows and 8 variables:
\describe{
  \item{icd10cm}{ICD-10-PCS procedure code}
  \item{icd9cm}{ICS-9-PCS procedure code}
  \item{flags}{(Describe "flags")}
  \item{approximate}{(Describe "approximate")}
  \item{no_map}{(Describe "no_map")}
  \item{combination}{(Describe "combination")}
  \item{scenario}{(Describe "scenario")}
  \item{choice_list}{(Describe "choice_list")}
}
}
\source{
\url{https://www.nber.org/research/data/icd-9-cm-and-icd-10-cm-and-icd-10-pcs-crosswalk-or-general-equivalence-mappings}
}
\usage{
icd10pcs_to_9
}
\description{
A dataset containing crosswalk values from ICD-10 procedure codes to ICD-9
procedure codes
}
\keyword{datasets}
