% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_funcs.R
\name{icd9_to_icd10}
\alias{icd9_to_icd10}
\title{Convert ICD-9-CM values to ICD-10-CM}
\usage{
icd9_to_icd10(icd_list)
}
\arguments{
\item{icd_list}{A vector of ICD-9 values}
}
\description{
`icd9_to_icd10` converts a vector of ICD-9 values to possible ICD-10 values.
Uses the NBER crosswalk file.
}
\examples{
# Get ICD-10 codes for CDI
icd9_to_icd10("00845")

# To collect all codes use children() from `icd` package
icd9_to_icd10("00845") \%>\% children()
}
