% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9pcs_to_10}
\alias{icd9pcs_to_10}
\title{NBER Procedure Crosswalk: ICD-9-PCS to ICD-10-PCS}
\format{
A data frame with 69,366 rows and 8 variables:
\describe{
  \item{icd9cm}{ICD-9-PCS procedure code}
  \item{icd10cm}{ICD-10-PCS procedure code}
  \item{flags}{(Describe "flags")}
  \item{approximate}{(Describe "approximate")}
  \item{no_map}{(Describe "no_map")}
  \item{combination}{(Describe "combination")}
  \item{scenario}{(Describe "scenario")}
  \item{choice_list}{(Describe "choice_list")}
}
}
\source{
\url{https://www.nber.org/research/data/icd-9-cm-and-icd-10-cm-and-icd-10-pcs-crosswalk-or-general-equivalence-mappings}
}
\usage{
icd9pcs_to_10
}
\description{
A dataset containing crosswalk values from ICD-9 procedure codes to ICD-10
procedure codes
}
\keyword{datasets}
