% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{load_ccs_codes}
\alias{load_ccs_codes}
\title{Load diagnosis codes for for a set of ccs codes}
\usage{
load_ccs_codes(ccs_codes)
}
\arguments{
\item{symptom_names}{a vector of project names to get codes for}
}
\description{
Return the sets of codes that correspond to particular CCS codes. This function
loads the icd-9 and icd-10 codes in the dx code list format along with the
description of the ccs category
}
