% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_funcs.R, R/load_codes.R
\name{load_disease_codes}
\alias{load_disease_codes}
\title{Load codes (diagnosis, procedure and medication) for particular diseases}
\usage{
load_disease_codes(project_names)

load_disease_codes(project_names)
}
\arguments{
\item{project_names}{a vector of project names to get codes for}
}
\description{
collect data from the core table

Return the sets of codes used to identify a particular disease
}
