% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{load_misc_codes}
\alias{load_misc_codes}
\title{Load misc code sets}
\usage{
load_misc_codes(misc_group)
}
\arguments{
\item{misc_group}{an misc code set to load}
}
\description{
Get miscellaneous code sets
}
