% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{load_procedure_codes}
\alias{load_procedure_codes}
\title{Load codes for a particular procedure}
\usage{
load_procedure_codes(procedure_names)
}
\arguments{
\item{procedure_names}{a vector of procedure names to get codes for}
}
\description{
Return the sets of codes used to identify a particular procedure
}
