% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook_funcs.R
\name{load_redbook}
\alias{load_redbook}
\title{Get Entire Redbook}
\usage{
load_redbook(redbook_path = "~/Data/redbook.csv")
}
\arguments{
\item{redbook_path}{The path to the redbook file.}
}
\description{
`load_redbook` returns the redbook file as a tibble
}
