% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{load_rx_codes}
\alias{load_rx_codes}
\title{Load ndc codes for particular group of medications}
\usage{
load_rx_codes(rx_group)
}
\arguments{
\item{rx_group}{an rx group to get ndc codes for}
}
\description{
Get NDC codes for a group of drugs
}
