% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_codes.R
\name{load_symptom_codes}
\alias{load_symptom_codes}
\title{Load diagnosis codes for particular symptoms}
\usage{
load_symptom_codes(symptom_names)
}
\arguments{
\item{symptoms_names}{a vector of project names to get codes for}
}
\description{
Return the sets of codes used to identify a particular symptoms
}
